Meter Readings, Payments, and Customer Creation offline (#19)
This PR allows meter readings, payments, and customers to be able to be created offline.

Workbox already allows Meter Readings / Payments to be created for customers that are already in the airtable. The main workflow that this PR addresses is adding meter readings / payments for a customer that isn't in the airtable yet (for example, the customer could've been created while offline and the user hasn't come back online yet).