Changed route names to be kebab-case and work with offline (#30)
Changed route names to be more CRUD-like (kebab-case and ending verbs).

Route have also been changed to not depend on rid. This is because users who are created offline do not receive an rid until they return online and are created on the Airtable. Therefore, we can't create routes based on rid because not all customers have it. Instead, because we store customer information anyways for offline use, we have a generic /customers/customer route that we can pass customer information to populate.